Revamped everything
Image conversion now uses an easily editable ini file for conversion.
On a per-group basis, sprites are quantized to find the closest matching
palette, and then remapped
Compression support is also implemented (might be a good idea to add a
decompression routine to the libraries so people can use it easier) --
LZ77 and RLE are considered the best choices here.
Anyways, it is a lot better and 100x easier to use.